requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
backports.entry-points-selectable==1.1.0
black==22.1.0
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.4.0
click==8.0.4
cryptography==41.0.3
distlib==0.3.2
dj-database-url==2.3.0
dj-email-url==1.0.6
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-cache-url==3.4.5
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.14.0
drf-jwt==1.19.2
environs==9.5.0
filelock==3.0.12
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.12.1
idna==3.10
imutils==0.5.4
keras==3.6.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
ml-dtypes==0.4.1
mypy-extensions==0.4.3
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.7.0.68
opt_einsum==3.4.0
optree==0.13.0
packaging==24.2
pathspec==0.9.0
pillow==11.0.0
pip-autoremove==0.10.0
platformdirs==2.2.0
protobuf==5.28.3
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2-binary==2.9.10
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2021.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
scipy==1.10.1
six==1.16.0
sqlparse==0.4.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.6.0
Werkzeug==3.1.3
whitenoise==6.4.0
wrapt==1.16.0
